9077: move processMessageEntries into its own file
