+++ shared/src/business/useCases/elasticsearch/processStreamUtilities.js
@@ -3,7 +3,7 @@ const {
   OPINION_EVENT_CODES_WITH_BENCH_OPINION,
   ORDER_EVENT_CODES,
 } = require('../../entities/EntityConstants');
-const { compact, flattenDeep, partition } = require('lodash');
+const { flattenDeep, partition } = require('lodash');
 const { processEntries } = require('./processEntries');
 
 const partitionRecords = records => {
@@ -177,71 +177,6 @@ const processWorkItemEntries = ({ applicationContext, workItemRecords }) =>
     records: workItemRecords,
   });
 
-const processMessageEntries = async ({
-  applicationContext,
-  messageRecords,
-}) => {
-  if (!messageRecords.length) return;
-
-  applicationContext.logger.debug(
-    `going to index ${messageRecords.length} message records`,
-  );
-
-  const indexMessageEntry = async messageRecord => {
-    const messageNewImage = messageRecord.dynamodb.NewImage;
-
-    // go get the latest message if we're indexing a message with isRepliedTo set to false - it might
-    // have been updated in dynamo since this record was created to be processed
-    if (!messageNewImage.isRepliedTo.BOOL) {
-      const latestMessageData = await applicationContext
-        .getPersistenceGateway()
-        .getMessage({
-          applicationContext,
-          docketNumber: messageNewImage.docketNumber.S,
-          messageId: messageNewImage.messageId.S,
-        });
-
-      if (!latestMessageData.isRepliedTo) {
-        const marshalledMessage =
-          AWS.DynamoDB.Converter.marshall(latestMessageData);
-
-        return {
-          dynamodb: {
-            Keys: {
-              pk: {
-                S: messageNewImage.pk.S,
-              },
-              sk: {
-                S: messageNewImage.sk.S,
-              },
-            },
-            NewImage: marshalledMessage,
-          },
-          eventName: 'MODIFY',
-        };
-      }
-    } else {
-      return messageRecord;
-    }
-  };
-
-  const indexRecords = await Promise.all(messageRecords.map(indexMessageEntry));
-
-  const { failedRecords } = await applicationContext
-    .getPersistenceGateway()
-    .bulkIndexRecords({
-      applicationContext,
-      records: compact(indexRecords),
-    });
-
-  if (failedRecords.length > 0) {
-    applicationContext.logger.error('the records that failed to index', {
-      failedRecords,
-    });
-    throw new Error('failed to index message records');
-  }
-};
-
 const practitionerEntityTypes = ['PrivatePractitioner', 'IrsPractitioner'];
 const practitionerSortKeys = ['privatePractitioner', 'irsPractitioner'];
 
@@ -370,7 +305,6 @@ const processOtherEntries = ({ applicationContext, otherRecords }) =>
 exports.partitionRecords = partitionRecords;
 exports.processPractitionerMappingEntries = processPractitionerMappingEntries;
 exports.processDocketEntries = processDocketEntries;
-exports.processMessageEntries = processMessageEntries;
 exports.processOtherEntries = processOtherEntries;
 exports.processRemoveEntries = processRemoveEntries;
 exports.processWorkItemEntries = processWorkItemEntries;
